--- proxy/convert/Cargo.toml
@@ -1,4 +1,0 @@
-[package]
-name = "convert"
-version = "0.3.0"
-publish = false
